refactor(auth, transactions): restructure auth & add transaction use cases
- migrate auth logic from legacy repository handlers to clean use case pattern with IAuthService
- replace old auth handlers with LoginUseCase and LogoutUseCase implementations
- add complete transaction CRUD use cases (create, delete, edit, fetch)
- update all presentation hooks to use the new use cases
- add unit tests for all new use cases and auth services
- remove deprecated auth handler files, tests, and IAuthRepository interface
- implement new logout page with loading state
