+++ src/application/services/iauth.service.ts
@@ -10,4 +10,14 @@ export interface IAuthService {
    * Returns the currently authenticated user or null if not authenticated.
    */
   getCurrentUser(): Promise<User | null>;
+
+  /**
+   * Signs in the user with Google OAuth.
+   */
+  signInWithGoogle(redirectTo: string): Promise<void>;
+
+  /**
+   * Signs out the currently authenticated user.
+   */
+  signOut(): Promise<void>;
 }
