+++ src/application/usecases/auth/login/iusecase.ts
@@ -1,0 +1,3 @@
+export interface ILoginUseCase {
+  loginWithGoogle: (redirectTo: string) => Promise<void>;
+}
