--- src/application/usecases/auth/logout.handler.ts
@@ -1,5 +1,0 @@
-import { authRepository } from '../../../infrastructure/supabase/auth.repository';
-
-export async function logoutHandler(): Promise<void> {
-  await authRepository.signOut();
-}
