+++ src/application/usecases/auth/logout/iusecase.ts
@@ -1,0 +1,3 @@
+export interface ILogoutUseCase {
+  execute(): Promise<void>;
+}
