+++ src/application/usecases/transaction/get-transaction/iusecase.ts
@@ -1,0 +1,7 @@
+import { Transaction } from "@/domain/entities/transaction/transaction";
+
+interface IGetTransactionsUseCase {
+  execute(monthYear: string): Promise<Transaction[]>;
+}
+
+export default IGetTransactionsUseCase;
