--- src/domain/repositories/IAuthRepository.ts
@@ -1,4 +1,0 @@
-export interface IAuthRepository {
-  signInWithGoogle(redirectTo: string): Promise<void>;
-  signOut(): Promise<void>;
-}
