+++ src/presentation/hooks/use-edit-transaction.test.ts
@@ -1,8 +1,8 @@
 import { renderHook, act } from '@testing-library/react';
 import { useEditTransaction } from './use-edit-transaction';
-import { EditTransactionUseCase } from '@/application/usecases/transaction/edit-transaction/edit-transaction.usecase';
+import { EditTransactionUseCase } from '@/application/usecases/transaction/edit-transaction/usecase';
 
-jest.mock('@/application/commands/transaction/edit-transaction/edit-transaction.usecase');
+jest.mock('@/application/usecases/transaction/edit-transaction/usecase');
 
 describe('useEditTransaction', () => {
   const mockInput = { amount: 200 } as any;
