+++ src/presentation/hooks/use-transactions.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from "react";
 import { Transaction } from "@/domain/entities/transaction/transaction";
-import { listTransactionsHandler } from "@/application/queries/list-transactions/list-transactions.handler";
+import { listTransactionsHandler } from '@/application/usecases/transaction/get-transaction/handler';
 
 export function useTransactions(monthYear: string) {
   const [transactions, setTransactions] = useState<Transaction[]>([]);
